ANDROID 55: Добавлен OnResult
