+++ feature/public-questions/impl/src/main/java/ru/yeahub/public_questions/impl/domain/usecase/GetPublicQuestionsUseCase.kt
@@ -2,7 +2,7 @@ package ru.yeahub.public_questions.impl.domain.usecase
 
 import ru.yeahub.public_questions.impl.domain.entity.PublicQuestionsModel
 import ru.yeahub.public_questions.impl.domain.repository.PublicQuestionsRepositoryApi
-import ru.yeahub.public_questions.impl.presentation.RequestPublicQuestionsData
+import ru.yeahub.public_questions.impl.presentation.viewmodel.RequestPublicQuestionsData
 
 class GetPublicQuestionsUseCase(private val repository: PublicQuestionsRepositoryApi) {
