+++ feature/public-questions/impl/src/main/java/ru/yeahub/public_questions/impl/presentation/intents/PublicQuestionsResult.kt
@@ -1,0 +1,6 @@
+package ru.yeahub.public_questions.impl.presentation.intents
+
+sealed class PublicQuestionsResult {
+    data object NavigateBack : PublicQuestionsResult()
+    data class NavigateToDetail(val id: String) : PublicQuestionsResult()
+}
