+++ feature/public-questions/impl/src/main/java/ru/yeahub/public_questions/impl/presentation/intents/PublicQuestionsScreenEvent.kt
@@ -1,4 +1,4 @@
-package ru.yeahub.public_questions.impl.presentation.event
+package ru.yeahub.public_questions.impl.presentation.intents
 
 sealed class PublicQuestionsScreenEvent {
